feat: Updating views & gen types

+++ web/bin/images.mjs
@@ -130,8 +130,22 @@ export const imageSizes = [
  */
 function generateFileName(filePath, sizeName, formatOptions) {
 	const originalName = path.parse(path.basename(filePath)).name;
+	const originalExtension = path.parse(filePath).ext;
 	const formattedSizeName = sizeName.replaceAll("_", "-");
-	const formatExtension = formatOptions ? `.${formatOptions.format}` : ".jpg";
+	const formatExtension = formatOptions ? `.${formatOptions.format}` : originalExtension;
+
+	if (formattedSizeName === 'webp' || formattedSizeName === 'avif') {
+		return `${originalName}${formatExtension}`;
+	}
+
+	if (formattedSizeName.includes('avif') || (formattedSizeName.includes('webp'))) {
+		const size = formattedSizeName
+			.replace('avif', '')
+			.replace('webp', '')
+			.replaceAll('-', '')
+		return `${originalName}-${size}${formatExtension}`
+	}
+
 	return `${originalName}-${formattedSizeName}${formatExtension}`;
 }
 
@@ -180,12 +194,26 @@ async function copyAndConvertImages(srcDir, destDir) {
 			const filePath = path.join(srcDir, file);
 			const destPathOriginal = path.join(destDir, file);
 
-			const fileStat = await fs.stat(filePath);
-			if (fileStat.isFile()) {
-				// Copy the original file
-				await fs.copy(filePath, destPathOriginal);
+			if (path.extname(file).toLowerCase() === '.svg') {
+				continue;
+			}
 
-				console.log(fileStat);
+			if (fs.exists(destPathOriginal)) {
+				continue
+			}
+
+			const fileStat = await fs.stat(filePath);
+			if (fileStat.isDirectory()) {
+				await copyAndConvertImages(filePath, destPathOriginal);
+			} else if (fileStat.isFile()) {
+
+				// Copy the original file
+				try {
+					await fs.copy(filePath, destPathOriginal);
+				} catch(err) {
+					console.log("hey", filePath, destPathOriginal, err);
+				}
+
 
 				// Process each size configuration
 				for (const size of imageSizes) {
